more work on the client refactor, i think all the previous functionality is already covered
