+++ client/src/ds2.cpp
@@ -1,0 +1,707 @@
+#include "ds2.h"
+
+// these were calculated by multiplying
+// the base price of the item with the price rate
+std::map<int, float> shop_prices = {
+    {30700000, 4300.0f},
+    {30700001, 5000.0f},
+    {30700002, 7000.0f},
+    {30700003, 7500.0f},
+    {30700100, 4500.0f},
+    {30700300, 2700.0f},
+    {30700600, 1500.0f},
+    {30700601, 400.0f},
+    {30700602, 180.0f},
+    {50600300, 12000.0f},
+    {50600301, 10400.0f},
+    {50600302, 8800.0f},
+    {50600303, 5200.0f},
+    {50600401, 9000.0f},
+    {50600600, 3000.0f},
+    {50600601, 4500.0f},
+    {50600602, 60.0f},
+    {50600603, 80.0f},
+    {70400000, 1600.0f},
+    {70400001, 1200.0f},
+    {70400002, 1440.0f},
+    {70400003, 960.0f},
+    {70400004, 2000.0f},
+    {70400005, 1500.0f},
+    {70400006, 1800.0f},
+    {70400007, 1200.0f},
+    {70400200, 800.0f},
+    {70400201, 960.0f},
+    {70400202, 800.0f},
+    {70400203, 880.0f},
+    {70400204, 1000.0f},
+    {70400205, 1200.0f},
+    {70400206, 1000.0f},
+    {70400207, 1100.0f},
+    {70400500, 6399.2f},
+    {70400501, 7999.0f},
+    {70400600, 1600.0f},
+    {70400601, 280.0f},
+    {70400602, 1600.0f},
+    {70400603, 2000.0f},
+    {70400604, 350.0f},
+    {70400605, 2000.0f},
+    {72110000, 4200.0f},
+    {72110001, 4800.0f},
+    {72110002, 8000.0f},
+    {72110003, 8000.0f},
+    {72110004, 8000.0f},
+    {72110005, 8000.0f},
+    {72110006, 8000.0f},
+    {72110007, 10000.0f},
+    {72110008, 10000.0f},
+    {72110100, 8000.0f},
+    {72110101, 8000.0f},
+    {72110102, 8000.0f},
+    {72110103, 8000.0f},
+    {72110104, 8000.0f},
+    {72110105, 10000.0f},
+    {72110300, 8200.0f},
+    {72110301, 11000.0f},
+    {72110302, 17400.0f},
+    {72110303, 9300.0f},
+    {72110400, 9000.0f},
+    {72110600, 2500.0f},
+    {72110601, 4000.0f},
+    {72110602, 6000.0f},
+    {72110603, 2500.0f},
+    {72110604, 60.0f},
+    {72110605, 60.0f},
+    {72110606, 80.0f},
+    {72110607, 80.0f},
+    {72500300, 3000.0f},
+    {72500301, 600.0f},
+    {72500302, 1500.0f},
+    {72500303, 4700.0f},
+    {72500304, 13000.0f},
+    {72500600, 3000.0f},
+    {72500601, 7500.0f},
+    {72600400, 5000.0f},
+    {72600401, 6000.0f},
+    {72600402, 7000.0f},
+    {72600600, 1500.0f},
+    {72600601, 2000.0f},
+    {72600602, 350.0f},
+    {72600603, 60.0f},
+    {72600604, 1500.0f},
+    {72600605, 2000.0f},
+    {72600606, 350.0f},
+    {72600607, 60.0f},
+    {75400000, 400.0f},
+    {75400001, 700.0f},
+    {75400002, 800.0f},
+    {75400003, 1000.0f},
+    {75400004, 15000.0f},
+    {75400005, 10000.0f},
+    {75400006, 8800.0f},
+    {75400007, 5000.0f},
+    {75400100, 5000.0f},
+    {75400101, 5800.0f},
+    {75400200, 850.0f},
+    {75400201, 1500.0f},
+    {75400202, 1000.0f},
+    {75400203, 1280.0f},
+    {75400204, 6000.0f},
+    {75400205, 4800.0f},
+    {75400206, 7000.0f},
+    {75400207, 5100.0f},
+    {75400208, 6200.0f},
+    {75400209, 5000.0f},
+    {75400210, 7400.0f},
+    {75400211, 5900.0f},
+    {75400212, 6400.0f},
+    {75400213, 3000.0f},
+    {75400214, 5500.0f},
+    {75400215, 3800.0f},
+    {75400216, 4600.0f},
+    {75400217, 5500.0f},
+    {75400218, 5200.0f},
+    {75400219, 6800.0f},
+    {75400220, 5200.0f},
+    {75400221, 6000.0f},
+    {75400222, 6000.0f},
+    {75400223, 4200.0f},
+    {75400224, 4800.0f},
+    {75400225, 3600.0f},
+    {75400226, 6000.0f},
+    {75400227, 4200.0f},
+    {75400228, 4800.0f},
+    {75400229, 3600.0f},
+    {75400230, 6000.0f},
+    {75400231, 4200.0f},
+    {75400232, 4800.0f},
+    {75400233, 6000.0f},
+    {75400234, 4200.0f},
+    {75400235, 4800.0f},
+    {75400236, 5000.0f},
+    {75400237, 4800.0f},
+    {75400238, 7200.0f},
+    {75400239, 5400.0f},
+    {75400240, 6300.0f},
+    {75400300, 1500.0f},
+    {75400301, 1500.0f},
+    {75400500, 1000.0f},
+    {75400600, 300.0f},
+    {75400601, 300.0f},
+    {75400602, 1000.0f},
+    {75400603, 1600.0f},
+    {75400604, 100.0f},
+    {75400605, 400.0f},
+    {75400606, 200.0f},
+    {75400607, 1500.0f},
+    {75400608, 4000.0f},
+    {75400609, 6000.0f},
+    {75400610, 12000.0f},
+    {75400611, 18000.0f},
+    {75400612, 24000.0f},
+    {75400613, 36000.0f},
+    {75400614, 48000.0f},
+    {75400615, 12000.0f},
+    {75600600, 1600.0f},
+    {75600601, 1600.0f},
+    {75600602, 2000.0f},
+    {75600603, 600.0f},
+    {75600604, 700.0f},
+    {76100000, 1200.0f},
+    {76100001, 1500.0f},
+    {76100002, 2000.0f},
+    {76100200, 800.0f},
+    {76100201, 1260.0f},
+    {76100202, 940.0f},
+    {76100203, 1100.0f},
+    {76100204, 900.0f},
+    {76100205, 1500.0f},
+    {76100206, 1050.0f},
+    {76100207, 1200.0f},
+    {76100208, 1000.0f},
+    {76100209, 750.0f},
+    {76100210, 850.0f},
+    {76100211, 1600.0f},
+    {76100212, 2700.0f},
+    {76100213, 1900.0f},
+    {76100214, 2200.0f},
+    {76100215, 2200.0f},
+    {76100216, 4000.0f},
+    {76100217, 2500.0f},
+    {76100218, 2900.0f},
+    {76100219, 1600.0f},
+    {76100220, 2700.0f},
+    {76100221, 1900.0f},
+    {76100222, 2200.0f},
+    {76100223, 2200.0f},
+    {76100224, 4000.0f},
+    {76100225, 2500.0f},
+    {76100226, 2900.0f},
+    {76100227, 5200.0f},
+    {76100228, 8000.0f},
+    {76100229, 6000.0f},
+    {76100230, 7100.0f},
+    {76100235, 7800.0f},
+    {76100236, 10000.0f},
+    {76100237, 8500.0f},
+    {76100238, 9000.0f},
+    {76100239, 8200.0f},
+    {76100240, 13500.0f},
+    {76100241, 9500.0f},
+    {76100242, 10800.0f},
+    {76100243, 7000.0f},
+    {76100244, 8200.0f},
+    {76100245, 6800.0f},
+    {76100246, 7500.0f},
+    {76100247, 9800.0f},
+    {76100248, 13000.0f},
+    {76100249, 10700.0f},
+    {76100250, 12000.0f},
+    {76100251, 11000.0f},
+    {76100252, 15000.0f},
+    {76100253, 12600.0f},
+    {76100254, 13800.0f},
+    {76100255, 9300.0f},
+    {76100256, 12500.0f},
+    {76100257, 10200.0f},
+    {76100258, 11600.0f},
+    {76100259, 10000.0f},
+    {76100260, 15000.0f},
+    {76100261, 10000.0f},
+    {76100262, 10000.0f},
+    {76100263, 14400.0f},
+    {76100264, 24000.0f},
+    {76100265, 16800.0f},
+    {76100266, 19200.0f},
+    {76200300, 8800.0f},
+    {76200301, 4000.0f},
+    {76200600, 1000.0f},
+    {76200601, 1000.0f},
+    {76200602, 2500.0f},
+    {76200603, 2500.0f},
+    {76200604, 6000.0f},
+    {76200605, 6000.0f},
+    {76200606, 13000.0f},
+    {76200607, 15000.0f},
+    {76200608, 17500.0f},
+    {76200609, 7000.0f},
+    {76200610, 7000.0f},
+    {76200611, 7000.0f},
+    {76200612, 7000.0f},
+    {76200613, 7000.0f},
+    {76200614, 7000.0f},
+    {76200615, 6000.0f},
+    {76200616, 7000.0f},
+    {76200617, 10000.0f},
+    {76200618, 7500.0f},
+    {76300300, 1200.0f},
+    {76300301, 3400.0f},
+    {76300302, 1500.0f},
+    {76300303, 3400.0f},
+    {76300304, 2300.0f},
+    {76300305, 3500.0f},
+    {76300400, 2000.0f},
+    {76300401, 2000.0f},
+    {76300402, 2000.0f},
+    {76300600, 1200.0f},
+    {76300601, 1200.0f},
+    {76300602, 1200.0f},
+    {76300603, 8000.0f},
+    {76400000, 1300.0f},
+    {76400001, 1700.0f},
+    {76400002, 1500.0f},
+    {76400003, 1000.0f},
+    {76400004, 1200.0f},
+    {76400005, 1500.0f},
+    {76400006, 1200.0f},
+    {76400007, 1400.0f},
+    {76400600, 10.0f},
+    {76400601, 30.0f},
+    {76400602, 25.0f},
+    {76400603, 50.0f},
+    {76400604, 800.0f},
+    {76400605, 2500.0f},
+    {76430000, 5000.0f},
+    {76430001, 3000.0f},
+    {76430002, 6200.0f},
+    {76430003, 2500.0f},
+    {76430004, 3500.0f},
+    {76430005, 2000.0f},
+    {76430006, 1500.0f},
+    {76430007, 5000.0f},
+    {76430008, 3000.0f},
+    {76430009, 6200.0f},
+    {76430010, 2500.0f},
+    {76430011, 3500.0f},
+    {76430012, 2000.0f},
+    {76430013, 1500.0f},
+    {76430100, 2400.0f},
+    {76430102, 2400.0f},
+    {76430600, 10.0f},
+    {76430601, 30.0f},
+    {76430602, 25.0f},
+    {76430603, 50.0f},
+    {76430604, 120.0f},
+    {76430605, 2500.0f},
+    {76430606, 2500.0f},
+    {76600000, 1000.0f},
+    {76600300, 1500.0f},
+    {76600301, 3000.0f},
+    {76600302, 2000.0f},
+    {76600303, 4500.0f},
+    {76600304, 1800.0f},
+    {76600305, 2600.0f},
+    {76600306, 2000.0f},
+    {76600307, 3000.0f},
+    {76600308, 1500.0f},
+    {76600309, 3000.0f},
+    {76600310, 2000.0f},
+    {76600311, 4500.0f},
+    {76600312, 1800.0f},
+    {76600313, 2600.0f},
+    {76600314, 2000.0f},
+    {76600315, 3000.0f},
+    {76600400, 2000.0f},
+    {76600401, 2000.0f},
+    {76600600, 1600.0f},
+    {76600601, 2400.0f},
+    {76600602, 13000.0f},
+    {76800000, 10000.0f},
+    {76800001, 7500.0f},
+    {76800200, 3600.0f},
+    {76800201, 9000.0f},
+    {76800202, 6300.0f},
+    {76800203, 7200.0f},
+    {76800300, 6500.0f},
+    {76800301, 4500.0f},
+    {76800302, 13000.0f},
+    {76800303, 2200.0f},
+    {76800304, 12400.0f},
+    {76800305, 6700.0f},
+    {76800306, 9500.0f},
+    {76800307, 600.0f},
+    {76800308, 1500.0f},
+    {76800309, 5200.0f},
+    {76800310, 11500.0f},
+    {76800311, 6300.0f},
+    {76800312, 3000.0f},
+    {76800400, 28000.0f},
+    {76800401, 2500.0f},
+    {76800402, 5000.0f},
+    {76800600, 3000.0f},
+    {76801000, 1500.0f},
+    {76801001, 1500.0f},
+    {76801002, 1500.0f},
+    {76801003, 1500.0f},
+    {76801004, 1500.0f},
+    {76801005, 1500.0f},
+    {76801006, 1500.0f},
+    {76801007, 1500.0f},
+    {76801008, 1500.0f},
+    {76801009, 1500.0f},
+    {76801010, 1500.0f},
+    {76801011, 1500.0f},
+    {76801012, 1500.0f},
+    {76801013, 1500.0f},
+    {76801014, 1500.0f},
+    {76801015, 1500.0f},
+    {76801016, 1500.0f},
+    {76801017, 1500.0f},
+    {76801100, 1500.0f},
+    {76801101, 1500.0f},
+    {76801102, 1500.0f},
+    {76801103, 1500.0f},
+    {76801104, 3000.0f},
+    {76801105, 5000.0f},
+    {76801106, 5000.0f},
+    {76801107, 10000.0f},
+    {76801108, 10000.0f},
+    {76801109, 5000.0f},
+    {76801110, 10000.0f},
+    {76801111, 45000.0f},
+    {76801200, 1500.0f},
+    {76801201, 1500.0f},
+    {76801202, 1500.0f},
+    {76801203, 1500.0f},
+    {76801204, 1500.0f},
+    {76801205, 1500.0f},
+    {76801206, 1500.0f},
+    {76801207, 1500.0f},
+    {76801208, 1500.0f},
+    {76801209, 1500.0f},
+    {76801210, 1500.0f},
+    {76801211, 1500.0f},
+    {76801212, 1500.0f},
+    {76801213, 1500.0f},
+    {76801214, 1500.0f},
+    {76801215, 1500.0f},
+    {76801216, 1500.0f},
+    {76801217, 1500.0f},
+    {76801300, 1500.0f},
+    {76801301, 1500.0f},
+    {76801302, 1500.0f},
+    {76801303, 1500.0f},
+    {76801304, 3000.0f},
+    {76801305, 5000.0f},
+    {76801306, 5000.0f},
+    {76900000, 1400.0f},
+    {76900300, 1500.0f},
+    {76900301, 3000.0f},
+    {76900302, 4500.0f},
+    {76900303, 3000.0f},
+    {76900304, 4500.0f},
+    {76900305, 2000.0f},
+    {76900306, 1800.0f},
+    {76900307, 6000.0f},
+    {76900308, 2400.0f},
+    {76900309, 3700.0f},
+    {76900310, 1500.0f},
+    {76900311, 3000.0f},
+    {76900312, 4500.0f},
+    {76900313, 3000.0f},
+    {76900314, 4500.0f},
+    {76900315, 2000.0f},
+    {76900316, 1800.0f},
+    {76900317, 6000.0f},
+    {76900318, 2400.0f},
+    {76900319, 3700.0f},
+    {76900400, 28000.0f},
+    {77000000, 4000.0f},
+    {77000001, 4000.0f},
+    {77000002, 7500.0f},
+    {77000100, 7000.0f},
+    {77000300, 600.0f},
+    {77000301, 2700.0f},
+    {77000302, 3400.0f},
+    {77000303, 4000.0f},
+    {77000304, 2700.0f},
+    {77000305, 1100.0f},
+    {77000306, 3400.0f},
+    {77000400, 6000.0f},
+    {77000600, 1500.0f},
+    {77100200, 3600.0f},
+    {77100201, 6000.0f},
+    {77100202, 4200.0f},
+    {77100203, 4800.0f},
+    {77100204, 7200.0f},
+    {77100205, 7700.0f},
+    {77100206, 5600.0f},
+    {77100207, 6500.0f},
+    {77100300, 6000.0f},
+    {77100301, 6300.0f},
+    {77100400, 8400.0f},
+    {77100500, 7500.0f},
+    {77100600, 3600.0f},
+    {77100601, 4200.0f},
+    {77100602, 180.0f},
+    {77100603, 180.0f},
+    {77100604, 180.0f},
+    {77200200, 1000.0f},
+    {77200201, 1600.0f},
+    {77200202, 1200.0f},
+    {77200203, 1300.0f},
+    {77200204, 2000.0f},
+    {77200205, 3000.0f},
+    {77200206, 2300.0f},
+    {77200207, 2600.0f},
+    {77200208, 7200.0f},
+    {77200209, 11500.0f},
+    {77200210, 8000.0f},
+    {77200211, 9300.0f},
+    {77200212, 7200.0f},
+    {77200213, 12000.0f},
+    {77200214, 8400.0f},
+    {77200215, 9600.0f},
+    {77200216, 14400.0f},
+    {77200217, 24000.0f},
+    {77200218, 16800.0f},
+    {77200219, 19200.0f},
+    {77200300, 1500.0f},
+    {77200301, 3000.0f},
+    {77200302, 2000.0f},
+    {77200303, 4500.0f},
+    {77200304, 1500.0f},
+    {77200305, 4700.0f},
+    {77200306, 4800.0f},
+    {77200400, 9000.0f},
+    {77200500, 2000.0f},
+    {77200501, 2000.0f},
+    {77200502, 2000.0f},
+    {77200503, 2000.0f},
+    {77200600, 1300.0f},
+    {77200601, 1500.0f},
+    {77200602, 300.0f},
+    {77200603, 900.0f},
+    {77200604, 2500.0f},
+    {77200605, 3000.0f},
+    {77200606, 7500.0f},
+    {77600000, 5800.0f},
+    {77600001, 5000.0f},
+    {77600002, 3000.0f},
+    {77600003, 5500.0f},
+    {77600004, 4000.0f},
+    {77600300, 6500.0f},
+    {77600301, 6000.0f},
+    {77600302, 4800.0f},
+    {77600600, 600.0f},
+    {77600601, 60.0f},
+    {77600602, 80.0f},
+    {77600603, 1300.0f},
+    {77600604, 1600.0f},
+    {77600605, 1300.0f},
+    {77600606, 2400.0f},
+    {77601000, 1500.0f},
+    {77601001, 5000.0f},
+    {77601002, 5000.0f},
+    {77601003, 5000.0f},
+    {77601004, 1500.0f},
+    {77601005, 1500.0f},
+    {77601006, 3000.0f},
+    {77601007, 3000.0f},
+    {77601008, 3000.0f},
+    {77601009, 10000.0f},
+    {77601010, 10000.0f},
+    {77601011, 10000.0f},
+    {77601012, 1500.0f},
+    {77601013, 1500.0f},
+    {77601014, 1500.0f},
+    {77601015, 10000.0f},
+    {77601016, 10000.0f},
+    {77601017, 10000.0f},
+    {77601018, 3000.0f},
+    {77601019, 3000.0f},
+    {77601020, 3000.0f},
+    {77601021, 3000.0f},
+    {77601022, 10000.0f},
+    {77601023, 10000.0f},
+    {77601024, 10000.0f},
+    {77601025, 10000.0f},
+    {77601026, 14000.0f},
+    {77601027, 24000.0f},
+    {77601028, 20000.0f},
+    {77601029, 8000.0f},
+    {77601030, 25000.0f},
+    {77601031, 18000.0f},
+    {77601032, 1500.0f},
+    {77601033, 8000.0f},
+    {77601034, 1200.0f},
+    {77601035, 3000.0f},
+    {77601036, 5000.0f},
+    {77601037, 12000.0f},
+    {77601100, 1500.0f},
+    {77601101, 5000.0f},
+    {77601102, 5000.0f},
+    {77601103, 5000.0f},
+    {77601104, 1500.0f},
+    {77601105, 1500.0f},
+    {77601106, 3000.0f},
+    {77601107, 3000.0f},
+    {77601108, 3000.0f},
+    {77601109, 10000.0f},
+    {77601110, 10000.0f},
+    {77601111, 10000.0f},
+    {77601112, 1500.0f},
+    {77601113, 1500.0f},
+    {77601114, 1500.0f},
+    {77601115, 10000.0f},
+    {77601116, 10000.0f},
+    {77601117, 10000.0f},
+    {77601118, 3000.0f},
+    {77601119, 3000.0f},
+    {77601120, 3000.0f},
+    {77601121, 3000.0f},
+    {77602000, 0.0f},
+    {77602001, 0.0f},
+    {77602002, 0.0f},
+    {77602003, 0.0f},
+    {77602004, 0.0f},
+    {77602005, 0.0f},
+    {77602006, 0.0f},
+    {77602007, 0.0f},
+    {77602008, 0.0f},
+    {77602009, 0.0f},
+    {77602010, 0.0f},
+    {77602011, 0.0f},
+    {77602012, 0.0f},
+    {77602013, 0.0f},
+    {77602014, 0.0f},
+    {77602015, 0.0f},
+    {77602016, 0.0f},
+    {77602017, 0.0f},
+    {77602018, 0.0f},
+    {77602019, 0.0f},
+    {77602020, 0.0f},
+    {77602021, 0.0f},
+    {77602022, 0.0f},
+    {77602023, 0.0f},
+    {77602024, 0.0f},
+    {77602025, 0.0f},
+    {77602026, 0.0f},
+    {77602027, 0.0f},
+    {77602028, 0.0f},
+    {77602029, 0.0f},
+    {77602030, 0.0f},
+    {77602031, 0.0f},
+    {77602032, 0.0f},
+    {77602033, 0.0f},
+    {77602034, 0.0f},
+    {77602035, 0.0f},
+    {77602036, 0.0f},
+    {77602037, 0.0f},
+    {77602100, 0.0f},
+    {77602101, 0.0f},
+    {77602102, 0.0f},
+    {77602103, 0.0f},
+    {77602104, 0.0f},
+    {77602105, 0.0f},
+    {77602106, 0.0f},
+    {77602107, 0.0f},
+    {77602108, 0.0f},
+    {77602109, 0.0f},
+    {77602110, 0.0f},
+    {77602111, 0.0f},
+    {77602112, 0.0f},
+    {77602113, 0.0f},
+    {77602114, 0.0f},
+    {77602115, 0.0f},
+    {77602116, 0.0f},
+    {77602117, 0.0f},
+    {77602118, 0.0f},
+    {77602119, 0.0f},
+    {77602120, 0.0f},
+    {77602121, 0.0f},
+    {77700200, 6000.0f},
+    {77700400, 13400.0f},
+    {77700401, 6700.0f},
+    {77700402, 5500.0f},
+    {77700403, 5800.0f},
+    {77700404, 4500.0f},
+    {77700600, 300.0f},
+    {77700601, 300.0f},
+    {77700602, 1600.0f},
+    {77700603, 600.0f},
+    {77700604, 300.0f},
+    {77700605, 300.0f},
+    {77700606, 1600.0f},
+    {77700607, 600.0f},
+    {78300000, 3000.0f},
+    {78300001, 4000.0f},
+    {78300200, 3600.0f},
+    {78300201, 6000.0f},
+    {78300202, 4200.0f},
+    {78300203, 4800.0f},
+    {78300204, 5400.0f},
+    {78300205, 7500.0f},
+    {78300206, 6000.0f},
+    {78300207, 6700.0f},
+    {78300300, 4500.0f},
+    {78300301, 4800.0f},
+    {78300302, 3800.0f},
+    {78300400, 20000.0f},
+    {78300600, 5000.0f},
+    {78300601, 1300.0f},
+    {78300602, 350.0f},
+    {78300603, 10000.0f},
+    {78300604, 5000.0f},
+    {78400200, 4500.0f},
+    {78400201, 5000.0f},
+    {78400202, 4600.0f},
+    {78400203, 4800.0f},
+    {78400300, 8000.0f},
+    {78400301, 3000.0f},
+    {78400302, 2000.0f},
+    {78400303, 1800.0f},
+    {78400304, 4200.0f},
+    {78400305, 3300.0f},
+    {78400306, 3500.0f},
+    {78400307, 2200.0f},
+    {78400308, 8000.0f},
+    {78400309, 3000.0f},
+    {78400310, 2000.0f},
+    {78400311, 1800.0f},
+    {78400312, 4200.0f},
+    {78400313, 3300.0f},
+    {78400314, 3500.0f},
+    {78400315, 2200.0f},
+    {78400400, 5500.0f},
+    {78400401, 7000.0f},
+    {78400402, 9000.0f},
+    {78400600, 1200.0f},
+    {78500000, 1600.0f},
+    {78500001, 2500.0f},
+    {78500300, 4200.0f},
+    {78500301, 3300.0f},
+    {78500600, 2200.0f},
+    {78500601, 300.0f},
+    {78500602, 7000.0f},
+    {78500603, 10000.0f},
+};
+
+bool is_shop_location(int32_t id)
+{
+    return shop_prices.contains(id);
+}
+
+float_t get_shop_price(int32_t id)
+{
+    return shop_prices[id];
+}
