+++ client/src/randomizer.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <map>
+#include <cstdint>
+
+void override_item_params(std::map<int32_t, int32_t> rewards);
